[V0 Deprecation] Remove `vllm.worker` and update according imports (#25901)
